requirements.txt
attrs==21.4.0
aws-cdk-lib==2.14.0
boto3==1.21.8
botocore==1.24.8
cattrs==1.10.0
constructs==10.0.73
iniconfig==1.1.1
jmespath==0.10.0
jsii==1.54.0
packaging==21.3
pluggy==1.0.0
publication==0.0.3
py==1.11.0
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
PyYAML==6.0
s3transfer==0.5.2
six==1.16.0
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.8
